+++ src/mundo.c
@@ -22,6 +22,21 @@ enum mundo_attr {
 };
 
 /* Constructor/Destructor */
+// Constructor de mundo
+int mundo_build2(struct mundo * a, struct mundo *f, int t){
+    // Establece el tamaño de ambos mundos
+    mundo_set_tam(a,t);
+    mundo_set_tam(f,t);
+
+    // Reserva memoria para el tablero y comprueba la reserva
+    if (mundo_alloc_tablero(a) == -1 || mundo_alloc_tablero(f)  == -1) {
+        perror("Error ");
+        return -1;
+    }
+
+    return 0;
+}
+
 // Reserva el mundo
 struct mundo *mundo_alloc(){
     struct mundo *m;
@@ -30,6 +45,35 @@ struct mundo *mundo_alloc(){
     m = (struct mundo *) malloc(sizeof(struct mundo));
     m->flags = 0;
 
+    return m;
+}
+
+// Reserva el mundo cuando se le pasa en un fichero
+struct mundo *mundo_falloc(char *filename){
+    FILE *fpest;
+    printf("Leyendo fichero de estado anterior %s...\n", filename);
+    // Abre el fichero de estado y comprueba la apertura
+    if (!(fpest = fopen(filename, "r"))) {
+        perror("Error al abrir fichero ");
+        printf("Configuración predeterminada\n");
+    }
+
+    struct mundo *m = mundo_alloc();
+
+    // Lee el tamaño
+    fread(&m->tamanio, sizeof(int), 1, fpest);
+    if (ferror(fpest))
+        perror("Error al leer");
+
+    // Reserva el tablero
+    mundo_alloc_tablero(m);
+
+    // Lee el tablero
+    fread(m->tablero, sizeof(int), m->tamanio, fpest);
+    if (ferror(fpest))
+        perror("Error al leer");
+
+    fclose(fpest);
     return m;
 }
 
@@ -137,4 +181,25 @@ struct mundo *mundo_clone(struct mundo *actual, struct mundo *futuro) {
     memcpy(actual, futuro, sizeof(struct mundo));
 
     return actual;
+}
+
+// Guarda el estado del mundo entre ejecuciones
+void mundo_save(struct mundo *m) {
+    // Abre el fichero de estado y comprueba la apertura
+    FILE *fp;
+    if (!(fp = fopen("state.txt", "w")))
+        perror("Error ");
+
+    // Guarda el tamaño
+    int tam = mundo_get_tam(m);
+    fwrite(&tam, sizeof(tam), 1, fp);
+    if (ferror(fp))
+        perror("Error al escribir");
+
+    // Guarda el tablero
+    fwrite(mundo_get_tablero(m), mundo_get_tam(m), sizeof(int), fp);
+    if (ferror(fp))
+        perror("Error al escribir");
+
+    fclose(fp);
 }